+++ sdk/servicebus/Azure.Messaging.ServiceBus/src/ServiceBusReceivedMessage.cs
@@ -3,6 +3,7 @@
 
 using System;
 using System.Collections.Generic;
+using System.Collections.Immutable;
 using System.Globalization;
 
 namespace Azure.Messaging.ServiceBus
@@ -153,7 +154,7 @@ namespace Azure.Messaging.ServiceBus
         /// byte, sbyte, char, short, ushort, int, uint, long, ulong, float, double, decimal,
         /// bool, Guid, string, Uri, DateTime, DateTimeOffset, TimeSpan
         /// </remarks>
-        public IDictionary<string, object> Properties => SentMessage.Properties;
+        public ImmutableDictionary<string, object> Properties => SentMessage.Properties.ToImmutableDictionary();
 
         /// <summary>
         /// User property key representing deadletter reason, when a message is received from a deadletter subqueue of an entity.