+++ sdk/extensions/Azure.Extensions.AspNetCore.Configuration.Secrets/CHANGELOG.md
@@ -1,6 +1,9 @@
 # Release History
 
-## 1.0.0-preview.2 (Unreleased)
+## 1.0.0-preview.3 (Unreleased)
+
+
+## 1.0.0-preview.2 (2020-05-05)
 
 - Package renamed to Azure.Extensions.AspNetCore.Configuration.Secrets